Move GetLeavesByRange tests
